+++ src/classes_body_ships.py
@@ -23,3 +23,16 @@ class Small_ship(Vehicle):
         # return True if collision with map occurs
             if map.get_tile_type(coord) == "water" or map.get_tile_type(coord) == "shallow": return False
             else: return True
+
+
+class Medium_ship(Small_ship):
+    path = MEDIUM_SHIP_PATH
+    number_of_frames = MEDIUM_SHIP_FRAMES
+    number_of_frames_in_sequence = MEDIUM_SHIP_FRAMES - 1
+
+    v_max = 0.4
+    acceleration = 0.1
+    turn_speed = 0.005
+
+    hit_box_radius = 35
+    base_HP = 600