+++ src/settings.py
@@ -9,33 +9,41 @@ WIN_WIDTH, WIN_HEIGHT = 1600, 900
 FRAMERATE = 60
 
 # sprites - units
-TANK_PATH = ["imgs","vehicles","tank.png"]
-TANK_FRAMES = 2
+# TANK_PATH = ["imgs","vehicles","tank.png"]
+# TANK_FRAMES = 2
+# land units
 LIGHT_TRACK_PATH = ["imgs","vehicles","light_track.bmp"]
 LIGHT_TRACK_FRAMES = 5
 MEDIUM_TRACK_PATH = ["imgs","vehicles","medium_track.bmp"]
 MEDIUM_TRACK_FRAMES = 5
 HEAVY_TRACK_PATH = ["imgs","vehicles","heavy_track.bmp"]
 HEAVY_TRACK_FRAMES = 7
+HEAVY_TRACK_BASIC_PATH = ["imgs","vehicles","heavy_track_basic.bmp"]
+HEAVY_TRACK_BASIC_FRAMES = 7
 ANT_PATH = ["imgs","vehicles","ant.bmp"]
 ANT_FRAMES = 5
 
+# ships
 SMALL_SHIP_PATH = ["imgs","vehicles","small_ship.bmp"]
 SMALL_SHIP_FRAMES = 4
+MEDIUM_SHIP_PATH = ["imgs","vehicles","medium_ship.bmp"]
+MEDIUM_SHIP_FRAMES = 4
 
-TURRET_PATH = ["imgs","vehicles","turret.png"]
-LIGHT_CANNON_PATH = ["imgs","vehicles","light_cannon.bmp"]
-MEDIUM_CANNON_PATH = ["imgs","vehicles","medium_cannon.bmp"]
-MINIGUN_PATH = ["imgs","vehicles","minigun.bmp"]
-PLANE_MINIGUN_PATH = ["imgs","vehicles","plane_minigun.bmp"]
-SIDE_CANNON_PATH = ["imgs","vehicles","side_cannon.bmp"]
-
+# planes
 PLANE_PATH = ["imgs","vehicles","plane.bmp"]
 PLANE_FRAMES = 4
 BOMBER_PATH = ["imgs","vehicles","bomber.bmp"]
 BOMBER_FRAMES = 4
 STRATEGIC_BOMBER_PATH = ["imgs","vehicles","strategic_bomber.bmp"]
 STRATEGIC_BOMBER_FRAMES = 4
+
+#weapons
+# TURRET_PATH = ["imgs","vehicles","turret.png"]
+LIGHT_CANNON_PATH = ["imgs","weapons","light_cannon.bmp"]
+MEDIUM_CANNON_PATH = ["imgs","weapons","medium_cannon.bmp"]
+MINIGUN_PATH = ["imgs","weapons","minigun.bmp"]
+PLANE_MINIGUN_PATH = ["imgs","weapons","plane_minigun.bmp"]
+SIDE_CANNON_PATH = ["imgs","weapons","side_cannon.bmp"]
 
 # sprites - map
 TREES_PATH = ["imgs","map","trees.bmp"]